[optimze] edit activity href [fix]remove  Col column boolean, fix breadCrumb key
